+++ src/main/java/com/softbistro/survey/question/controller/QuestionController.java
@@ -1,15 +1,21 @@
 package com.softbistro.survey.question.controller;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.softbistro.survey.client.auth.service.AuthorizationService;
 import com.softbistro.survey.question.components.entity.Question;
 import com.softbistro.survey.question.service.QuestionService;
-import com.softbistro.survey.response.Response;
+
+import io.swagger.annotations.ApiOperation;
 
 /**
  * Controller for CRUD of Client
@@ -21,8 +27,13 @@ import com.softbistro.survey.response.Response;
 @RequestMapping(value = "/rest/survey/v1/question")
 public class QuestionController {
 
+	private static final Logger LOGGER = Logger.getLogger(QuestionController.class);
+
 	@Autowired
 	private QuestionService questionService;
+
+	@Autowired
+	private AuthorizationService authorizationService;
 
 	/**
 	 * Find question in database by id of question
@@ -31,10 +42,23 @@ public class QuestionController {
 	 *            id - id of question
 	 * @return return - all information about question
 	 */
+	@ApiOperation(value = "Get Question By Id", notes = "Get question instanse by question id", tags = "Question")
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public Response findQuestionById(@PathVariable("id") Long id) {
+	public ResponseEntity<Question> findQuestionById(@PathVariable("id") Integer id, @RequestHeader String token) {
 
-		return questionService.findQuestionById(id);
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
+		try {
+
+			return new ResponseEntity<>(questionService.findQuestionById(id), HttpStatus.OK);
+		} catch (Exception e) {
+
+			LOGGER.debug(e.getMessage());
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 	/**
@@ -45,10 +69,25 @@ public class QuestionController {
 	 *            database
 	 * @return return - status of execution this method
 	 */
+	@ApiOperation(value = "Create new Question", notes = "Create question instanse by survey id, question, description short,"
+			+ " description Long, questionSection Id, answer Type, question Choices, required, required Comment. "
+			+ "Field answerType can have value such as: RATE1-10, RATE1-5, RATE1-3, BOOLEAN, LIST, MULTILIST, INPUT, MEMO", tags = "Question")
 	@RequestMapping(value = "/", method = RequestMethod.POST)
-	public Response saveQuestion(@RequestBody Question question) {
+	public ResponseEntity<Object> saveQuestion(@RequestBody Question question, @RequestHeader String token) {
 
-		return questionService.saveQuestion(question);
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
+		try {
+
+			return new ResponseEntity<>(questionService.saveQuestion(question), HttpStatus.CREATED);
+		} catch (Exception e) {
+
+			LOGGER.debug(e.getMessage());
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 	/**
@@ -58,10 +97,24 @@ public class QuestionController {
 	 *            of question
 	 * @return return - status of execution this method
 	 */
+	@ApiOperation(value = "Delete Question By Id", notes = "Delete question instanse by question id", tags = "Question")
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public Response deleteQuestion(@PathVariable("id") Long id) {
+	public ResponseEntity<Object> deleteQuestion(@PathVariable("id") Integer id, @RequestHeader String token) {
 
-		return questionService.deleteQuestion(id);
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
+		try {
+
+			questionService.deleteQuestion(id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		} catch (Exception e) {
+
+			LOGGER.debug(e.getMessage());
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
 
 	/**
@@ -74,10 +127,26 @@ public class QuestionController {
 	 *            id-id of question
 	 * @return return - status of execution this method
 	 */
+	@ApiOperation(value = "Update Question By Id", notes = "Update question instanse by survey id, question, description short,"
+			+ " description Long, questionSection Id, answer Type, question Choices, required, required Comment and question id. "
+			+ "Field answerType can have value such as: RATE1-10, RATE1-5, RATE1-3, BOOLEAN, LIST, MULTILIST, INPUT, MEMO", tags = "Question")
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT)
-	public Response updateQuestion(@RequestBody Question question, @PathVariable("id") Long id) {
+	public ResponseEntity<Object> updateQuestion(@RequestBody Question question, @PathVariable("id") Integer id,
+			@RequestHeader String token) {
 
-		return questionService.updateQuestion(question, id);
+		if (!authorizationService.checkAccess(token)) {
+
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
+
+		try {
+
+			questionService.updateQuestion(question, id);
+			return new ResponseEntity<>(HttpStatus.OK);
+		} catch (Exception e) {
+
+			LOGGER.debug(e.getMessage());
+			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 	}
-
 }
